Fix pla purchased flag regression
ty cerquami for checking

+++ PKHeX.Core/Legality/Verifiers/LegendsArceusVerifier.cs
@@ -224,7 +224,7 @@ public sealed class LegendsArceusVerifier : Verifier
         // Check if the move can be learned in the learnset...
         // Changing forms do not have separate tutor permissions, so we don't need to bother with form changes.
         // Level up movepools can grant moves for mastery at lower levels for earlier evolutions... find the minimum.
-        int level = 101;
+        byte level = 101;
         foreach (var evo in data.Info.EvoChainsAllGens.Gen8a)
         {
             var moveset = LearnSource8LA.Instance.GetLearnset(evo.Species, evo.Form);
